+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unpublished
+### Added
+- add an example in `examples` directory
+
+
+
 ## 0.0.6 - 2020-01-11
 ### Changed
 - remove extra logs